% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adj}
\alias{adj}
\alias{gcv}
\title{Adjusted R-square statistics}
\usage{
adj(R2, n, p)

gcv(R2, n, p)
}
\arguments{
\item{R2}{a vector of R-square statistics}

\item{n}{length of the response}

\item{p}{number of columns of the response matrix}
}
\value{
an adjusted version of \code{R2}
}
\description{
Various adjustments to sample R-square statistics.  Currently only
adjusted and generalised cross validation adjustments are
available.  TODO:  add final prediction error
}

