% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bootCoef}
\alias{bootCoef}
\title{Bootstraped linear model coefficient}
\usage{
bootCoef(X, y, nBoot)
}
\arguments{
\item{X}{model matrix}

\item{y}{response vector}

\item{nBoot}{number of bootstrap samples}
}
\description{
Bootstraped linear model coefficient
}

