% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rUnif}
\alias{bootPerm}
\alias{rUnif}
\alias{shuffleMatrix}
\alias{shuffleVector}
\title{Uniform random number generation}
\usage{
rUnif(n)

bootPerm(n)

shuffleMatrix(X, prm)

shuffleVector(y, prm)
}
\arguments{
\item{n}{length of the permutation vector}

\item{X}{model matrix to be permuted}

\item{prm}{permutation of the observations (typically the output
of \code{bootPerm}}

\item{y}{response vector to be permuted}
}
\value{
a permutation of \code{1:n}
}
\description{
Uniform random number generation

Random permutation
}

