\name{bootR2-package}
\alias{bootR2-package}
\alias{bootR2}
\docType{package}
\title{
Predictive R-squares via the bootstrap
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab bootR2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-09\cr
License: \tab What Licence is it under ?\cr
}
}
\author{
Steve Walker
Maintainer: Steve Walker <steve.walker@utoronto.ca>
}
\keyword{ package }
