% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bootR2}
\alias{bootR2}
\alias{bootR2-package}
\alias{bootR2.default}
\alias{bootR2.lm}
\alias{bootR2.matrix}
\alias{bootR2.numeric}
\alias{bootR2pred}
\alias{bootR2samp}
\title{Bootstrap R-squared}
\usage{
bootR2(object, ...)

\method{bootR2}{default}(object, ...)

\method{bootR2}{lm}(object, nBoot = 1, predictive = TRUE, ...)

\method{bootR2}{matrix}(object, y, nBoot = 1, predictive = TRUE, ...)

\method{bootR2}{numeric}(object, X, nBoot = 1, predictive = TRUE, ...)

bootR2pred(X, y, nBoot)

bootR2samp(X, y, nBoot)
}
\arguments{
\item{object}{Object for calculating R-squared}

\item{...}{Not currently used}

\item{nBoot}{Number of bootstrap samples}

\item{predictive}{Should the predictive (i.e. double) bootstrap be
used?  If \code{TRUE}, each of the \code{nBoot} bootstrap samples
actually consists of two samples: one for training the linear
model and one for validating it.  Therefore, in this case, the
bootstrap distribution may be negative when out-of-sample
predictions generate too much variance relative to the bias they
avoid.  If \code{FALSE}, the same sample is used for both training
and validation.}

\item{X}{model matrix}

\item{y}{model response}
}
\value{
Vector of R-squared bootstrap samples
}
\description{
A package for computing fast Rcpp-based predictive R-squares, via
the nonparametric bootstrap.  Heavy use is made of RcppEigen.
Predictive R-squared statistics use two different samples, one for
training a linear model and one for validating it.  The
\code{bootR2} package draws two nonparametric bootstrap samples to
be used as training and validation data for calculating a
predictive R-squared. Such a bootstrap distribution provides a
better assessment of the proportion of variation that the fitted
model could explain in other samples from the population.  NOTE:
the adjusted R-squared does *not* provide this assessment.
}

