% Generated by roxygen2 (4.0.1): do not edit by hand
\name{iterSimExperiment}
\alias{iterSimExperiment}
\title{Predictive R-square simulations}
\usage{
iterSimExperiment(X, Y, pNoi, pSig, nIter, n)
}
\arguments{
\item{X}{model matrix in a population (last \code{pNoi} columns
should be unrelated to the model response)}

\item{Y}{model response in a population}

\item{pNoi}{number of predictors that are not related to the response}

\item{pSig}{number of predictors that are related to the response}

\item{nIter}{number of iterations}

\item{n}{sample size}
}
\value{
Two column matrix with estimated R-squares in the first
column and sample R-squares in the second
}
\description{
Calculate estimated and sample R-square statistics for each model
with zero to \code{pNoi} predictors that are unrelated to the
response
}

