% Generated by roxygen2 (4.0.1): do not edit by hand
\name{betaHat}
\alias{R2}
\alias{R2pred}
\alias{betaHat}
\title{Regression statistics}
\usage{
betaHat(X, y)

R2(X, y)

R2pred(Xt, yt, Xv, yv)
}
\arguments{
\item{X}{model matrix}

\item{y}{response vector (or matrix)}

\item{Xt}{training model matrix}

\item{yt}{training response vector}

\item{Xv}{validation model matrix}

\item{yv}{validation response vector}
}
\description{
\code{betaHat} returns the Linear model coefficient, \code{R2}
returns the R-squared statistic, and \code{R2pred} returns the
predictive R-squared statistic
}

